refactor(api): extract pages router (pure move)
Move the 16 pages-domain routes (/pages CRUD, current-display,
share/import, staff-picks, preview, cache, send) from api_server.py to
src/pages/routes.py, included at the exact spot the block occupied so
route registration order is unchanged. Handler bodies are verbatim;
only imports were adjusted.

Names still owned by api_server (service getters, get_service, the
board/silence/pause guards, text_to_board_array/resolve_dimensions/
is_collection_id) are imported inside each handler so they resolve
through src.api_server at call time — the suite patches them as
src.api_server.<name> (151x get_settings_service, 80x get_service, ...)
and a module-level import would detach the moved handlers from those
patches (and cycle). Same pattern src/mqtt/commands.py already uses.

_STAFF_PICKS_PATH gains one .parent (file is a directory deeper).
/pages/ai/* deliberately stays: it is AI-domain surface sharing the AI
semaphore/throttle helpers with /ai/* and moves with that domain later.

Response-shape goldens: zero changes. Route-inventory golden: zero
changes. pyproject gains a per-file lint-ignore entry mirroring
api_server's so the verbatim code lints clean; cleanup lands with #1766.

Part of #1849.
