+++ pyproject.toml
@@ -166,6 +166,10 @@ ignore = [
     "PTH",  # CLI scripts use os.path freely
     "T201", # allow print()
 ]
+# Routers extracted verbatim from src/api_server.py (issue #1756). They keep
+# api_server's lint-ignore set so the extraction stays a pure move; the
+# convention cleanup happens with #1766.
+"src/pages/routes.py" = ["PTH", "SIM", "RET", "PIE", "RUF010", "RUF013", "RUF059", "RUF100"]
 # CodeQL-sensitive files: see the matching [tool.ruff.format].exclude entry
 # above for the rationale. Any rule fix here would shift line numbers and
 # re-trigger CodeQL's "new alert" false positives on pre-existing findings.
